+++ packages/codemod/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vapor-ui/codemod
 
+## 1.0.0-beta.6
+
+### Minor Changes
+
+- b2d053c: add checkbox, collapsible, dialog transform
+- 939c845: add avatar, breadcrumb, card transform
+- eb3fdc7: add codemod cli packages
+
 ## 1.0.0-beta.5
 
 ### Minor Changes